Change strncasecmp() definition to match Bill's prototype. Fixes compile break on OS/2.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@85022 13f79535-47bb-0310-9956-ffa450edef68

+++ server/util.c
@@ -1713,7 +1713,7 @@ int strcasecmp(const char *a, const char *b)
 #endif
 
 #ifndef HAVE_STRNCASECMP
-int strncasecmp(const char *a, const char *b, int n)
+int strncasecmp(const char *a, const char *b, size_t n)
 {
     const char *p = a;
     const char *q = b;
